+++ mton/src/main/java/org/zerock/b01/service/warehouse/DashboardService.java
@@ -11,7 +11,8 @@ public interface DashboardService {
 
   Long getTodayPlannedCount();
 
-  int getCurrentMonthOrderTotal();
+//  int getCurrentMonthOrderTotal();
+  Long getCurrentMonthOrderTotal();
 
   double getCurrentMonthOrderProgressRate();
